[Decomp][Spell] Cohesion split: extract SpellEffectTail handlers to SpellEffectTail.cpp
