+++ src/game/WorldHandlers/SpellEffects.cpp
@@ -471,613 +471,21 @@ void Spell::EffectAddHonor(SpellEffectIndex /*eff_idx*/)
 
 
 
-/**
- * @brief Removes auras from the target that match the specified mechanic.
- *
- * @param eff_idx The effect index containing the mechanic id.
- */
-void Spell::EffectDispelMechanic(SpellEffectIndex eff_idx)
-{
-    if (!unitTarget)
-    {
-        return;
-    }
 
-    uint32 mechanic = m_spellInfo->EffectMiscValue[eff_idx];
 
-    Unit::SpellAuraHolderMap& Auras = unitTarget->GetSpellAuraHolderMap();
-    for (Unit::SpellAuraHolderMap::iterator iter = Auras.begin(), next; iter != Auras.end(); iter = next)
-    {
-        next = iter;
-        ++next;
-        SpellEntry const* spell = iter->second->GetSpellProto();
-        if (iter->second->HasMechanic(mechanic))
-        {
-            unitTarget->RemoveAurasDueToSpell(spell->Id);
-            if (Auras.empty())
-            {
-                break;
-            }
-            else
-            {
-                next = Auras.begin();
-            }
-        }
-    }
-}
 
-/**
- * @brief Restores the caster's dead pet and revives it with percentage-based health.
- */
-void Spell::EffectSummonDeadPet(SpellEffectIndex /*eff_idx*/)
-{
-    if (m_caster->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-    Player* _player = (Player*)m_caster;
-    Pet* pet = _player->GetPet();
-    if (!pet)
-    {
-        return;
-    }
-    if (pet->IsAlive())
-    {
-        return;
-    }
-    if (damage < 0)
-    {
-        return;
-    }
 
-    pet->SetUInt32Value(UNIT_DYNAMIC_FLAGS, UNIT_DYNFLAG_NONE);
-    pet->RemoveFlag(UNIT_FIELD_FLAGS, UNIT_FLAG_SKINNABLE);
-    pet->SetDeathState(ALIVE);
-    pet->clearUnitState(UNIT_STAT_ALL_STATE);
-    pet->SetHealth(uint32(pet->GetMaxHealth() * (float(damage) / 100)));
 
-    pet->AIM_Initialize();
 
-    // _player->PetSpellInitialize(); -- action bar not removed at death and not required send at revive
-    pet->SavePetToDB(PET_SAVE_AS_CURRENT);
-}
 
-/**
- * @brief Unsummons all totems currently owned by the caster.
- */
-void Spell::EffectDestroyAllTotems(SpellEffectIndex /*eff_idx*/)
-{
-    int32 mana = 0;
-    for (int slot = 0;  slot < MAX_TOTEM_SLOT; ++slot)
-    {
-        if (Totem* totem = m_caster->GetTotem(TotemSlot(slot)))
-        {
-            if (damage)
-            {
-                uint32 spell_id = totem->GetUInt32Value(UNIT_CREATED_BY_SPELL);
-                if (SpellEntry const* spellInfo = sSpellStore.LookupEntry(spell_id))
-                {
-                    uint32 manacost = spellInfo->manaCost + m_caster->GetCreateMana() * spellInfo->ManaCostPercentage / 100;
-                    mana += manacost * damage / 100;
-                }
-            }
-            {
-                totem->UnSummon();
-            }
-        }
-    }
 
-    if (mana)
-    {
-        m_caster->CastCustomSpell(m_caster, 39104, &mana, NULL, NULL, true);
-    }
-}
 
-/**
- * @brief Removes a fixed number of durability points from one or more player items.
- *
- * @param eff_idx The effect index containing the inventory slot selector.
- */
-void Spell::EffectDurabilityDamage(SpellEffectIndex eff_idx)
-{
-    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
 
-    int32 slot = m_spellInfo->EffectMiscValue[eff_idx];
 
-    // FIXME: some spells effects have value -1/-2
-    // Possibly its mean -1 all player equipped items and -2 all items
-    if (slot < 0)
-    {
-        ((Player*)unitTarget)->DurabilityPointsLossAll(damage, (slot < -1));
-        return;
-    }
 
-    // invalid slot value
-    if (slot >= INVENTORY_SLOT_BAG_END)
-    {
-        return;
-    }
 
-    if (Item* item = ((Player*)unitTarget)->GetItemByPos(INVENTORY_SLOT_BAG_0, slot))
-    {
-        ((Player*)unitTarget)->DurabilityPointsLoss(item, damage);
-    }
-}
 
-/**
- * @brief Removes a percentage of durability from one or more player items.
- *
- * @param eff_idx The effect index containing the inventory slot selector.
- */
-void Spell::EffectDurabilityDamagePCT(SpellEffectIndex eff_idx)
-{
-    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
 
-    int32 slot = m_spellInfo->EffectMiscValue[eff_idx];
 
-    // FIXME: some spells effects have value -1/-2
-    // Possibly its mean -1 all player equipped items and -2 all items
-    if (slot < 0)
-    {
-        ((Player*)unitTarget)->DurabilityLossAll(double(damage) / 100.0f, (slot < -1));
-        return;
-    }
 
-    // invalid slot value
-    if (slot >= INVENTORY_SLOT_BAG_END)
-    {
-        return;
-    }
 
-    if (damage <= 0)
-    {
-        return;
-    }
-
-    if (Item* item = ((Player*)unitTarget)->GetItemByPos(INVENTORY_SLOT_BAG_0, slot))
-    {
-        ((Player*)unitTarget)->DurabilityLoss(item, double(damage) / 100.0f);
-    }
-}
-
-/**
- * @brief Modifies the caster's threat on the target by a percentage.
- */
-void Spell::EffectModifyThreatPercent(SpellEffectIndex /*eff_idx*/)
-{
-    if (!unitTarget)
-    {
-        return;
-    }
-
-    unitTarget->GetThreatManager().modifyThreatPercent(m_caster, damage);
-}
-
-/**
- * @brief Summons a transmitted game object such as fishing nodes, rituals, or spell casters.
- *
- * @param eff_idx The effect index containing the game object entry.
- */
-void Spell::EffectTransmitted(SpellEffectIndex eff_idx)
-{
-    uint32 name_id = m_spellInfo->EffectMiscValue[eff_idx];
-
-    GameObjectInfo const* goinfo = ObjectMgr::GetGameObjectInfo(name_id);
-
-    if (!goinfo)
-    {
-        sLog.outErrorDb("Gameobject (Entry: %u) not exist and not created at spell (ID: %u) cast", name_id, m_spellInfo->Id);
-        return;
-    }
-
-    float fx, fy, fz;
-
-    if (m_targets.m_targetMask & TARGET_FLAG_DEST_LOCATION)
-    {
-        m_targets.getDestination(fx, fy, fz);
-    }
-    // FIXME: this can be better check for most objects but still hack
-    else if (m_spellInfo->EffectRadiusIndex[eff_idx] && m_spellInfo->speed == 0)
-    {
-        float dis = GetSpellRadius(sSpellRadiusStore.LookupEntry(m_spellInfo->EffectRadiusIndex[eff_idx]));
-        m_caster->GetClosePoint(fx, fy, fz, DEFAULT_WORLD_OBJECT_SIZE, dis);
-    }
-    else
-    {
-        float min_dis = GetSpellMinRange(sSpellRangeStore.LookupEntry(m_spellInfo->rangeIndex));
-        float max_dis = GetSpellMaxRange(sSpellRangeStore.LookupEntry(m_spellInfo->rangeIndex));
-        float dis = rand_norm_f() * (max_dis - min_dis) + min_dis;
-
-        // special code for fishing bobber (TARGET_SELF_FISHING), should not try to avoid objects
-        // nor try to find ground level, but randomly vary in angle
-        if (goinfo->type == GAMEOBJECT_TYPE_FISHINGNODE)
-        {
-            // calculate angle variation for roughly equal dimensions of target area
-            float max_angle = (max_dis - min_dis) / (max_dis + m_caster->GetObjectBoundingRadius());
-            float angle_offset = max_angle * (rand_norm_f() - 0.5f);
-            m_caster->GetNearPoint2D(fx, fy, dis + m_caster->GetObjectBoundingRadius(), m_caster->GetOrientation() + angle_offset);
-
-            GridMapLiquidData liqData;
-            if (!m_caster->GetTerrain()->IsInWater(fx, fy, m_caster->GetPositionZ() + 1.f, &liqData))
-            {
-                SendCastResult(SPELL_FAILED_NOT_FISHABLE);
-                SendChannelUpdate(0);
-                return;
-            }
-
-            fz = liqData.level;
-            // finally, check LoS
-            if (!m_caster->IsWithinLOS(fx, fy, fz))
-            {
-                SendCastResult(SPELL_FAILED_LINE_OF_SIGHT);
-                SendChannelUpdate(0);
-                return;
-            }
-        }
-        else
-        {
-            m_caster->GetClosePoint(fx, fy, fz, DEFAULT_WORLD_OBJECT_SIZE, dis);
-        }
-    }
-
-    Map* cMap = m_caster->GetMap();
-
-
-    // if gameobject is summoning object, it should be spawned right on caster's position
-    if (goinfo->type == GAMEOBJECT_TYPE_SUMMONING_RITUAL)
-    {
-        m_caster->GetPosition(fx, fy, fz);
-    }
-
-    GameObject* pGameObj = new GameObject;
-
-    if (!pGameObj->Create(cMap->GenerateLocalLowGuid(HIGHGUID_GAMEOBJECT), name_id, cMap,
-                          fx, fy, fz, m_caster->GetOrientation()))
-    {
-        delete pGameObj;
-        return;
-    }
-
-    int32 duration = GetSpellDuration(m_spellInfo);
-
-    switch (goinfo->type)
-    {
-        case GAMEOBJECT_TYPE_FISHINGNODE:
-        {
-            m_caster->SetChannelObjectGuid(pGameObj->GetObjectGuid());
-            m_caster->AddGameObject(pGameObj);              // will removed at spell cancel
-
-            // end time of range when possible catch fish (FISHING_BOBBER_READY_TIME..GetDuration(m_spellInfo))
-            // start time == fish-FISHING_BOBBER_READY_TIME (0..GetDuration(m_spellInfo)-FISHING_BOBBER_READY_TIME)
-            int32 lastSec = 0;
-            switch (urand(0, 3))
-            {
-                case 0: lastSec =  3; break;
-                case 1: lastSec =  7; break;
-                case 2: lastSec = 13; break;
-                case 3: lastSec = 17; break;
-            }
-
-            duration = duration - lastSec * IN_MILLISECONDS + FISHING_BOBBER_READY_TIME * IN_MILLISECONDS;
-            break;
-        }
-        case GAMEOBJECT_TYPE_SUMMONING_RITUAL:
-        {
-            if (m_caster->GetTypeId() == TYPEID_PLAYER)
-            {
-                pGameObj->AddUniqueUse((Player*)m_caster);
-                m_caster->AddGameObject(pGameObj);          // will removed at spell cancel
-            }
-            break;
-        }
-        case GAMEOBJECT_TYPE_FISHINGHOLE:
-        case GAMEOBJECT_TYPE_CHEST:
-        default:
-            break;
-    }
-
-    pGameObj->SetRespawnTime(duration > 0 ? duration / IN_MILLISECONDS : 0);
-
-    pGameObj->SetOwnerGuid(m_caster->GetObjectGuid());
-
-    pGameObj->SetUInt32Value(GAMEOBJECT_LEVEL, m_caster->getLevel());
-    pGameObj->SetSpellId(m_spellInfo->Id);
-
-    DEBUG_LOG("AddObject at SpellEfects.cpp EffectTransmitted");
-    // m_caster->AddGameObject(pGameObj);
-    // m_ObjToDel.push_back(pGameObj);
-
-    cMap->Add(pGameObj);
-    pGameObj->AIM_Initialize();
-
-    pGameObj->SummonLinkedTrapIfAny();
-
-    if (m_caster->GetTypeId() == TYPEID_UNIT && ((Creature*)m_caster)->AI())
-    {
-        ((Creature*)m_caster)->AI()->JustSummoned(pGameObj);
-    }
-    if (m_originalCaster && m_originalCaster != m_caster && m_originalCaster->GetTypeId() == TYPEID_UNIT && ((Creature*)m_originalCaster)->AI())
-    {
-        ((Creature*)m_originalCaster)->AI()->JustSummoned(pGameObj);
-    }
-}
-
-void Spell::EffectProspecting(SpellEffectIndex /*eff_idx*/)
-{
-    if (m_caster->GetTypeId() != TYPEID_PLAYER || !itemTarget)
-    {
-        return;
-    }
-
-    Player* p_caster = (Player*)m_caster;
-
-    if (sWorld.getConfig(CONFIG_BOOL_SKILL_PROSPECTING))
-    {
-        uint32 SkillValue = p_caster->GetPureSkillValue(SKILL_JEWELCRAFTING);
-        uint32 reqSkillValue = itemTarget->GetProto()->RequiredSkillRank;
-        p_caster->UpdateGatherSkill(SKILL_JEWELCRAFTING, SkillValue, reqSkillValue);
-    }
-
-    ((Player*)m_caster)->SendLoot(itemTarget->GetObjectGuid(), LOOT_PROSPECTING);
-}
-
-void Spell::EffectSkill(SpellEffectIndex /*eff_idx*/)
-{
-    DEBUG_LOG("WORLD: SkillEFFECT");
-}
-
-/**
- * @brief Fully resurrects a dead player target as part of a spirit heal effect.
- */
-void Spell::EffectSpiritHeal(SpellEffectIndex /*eff_idx*/)
-{
-    // TODO player can't see the heal-animation - he should respawn some ticks later
-    if (!unitTarget || unitTarget->IsAlive())
-    {
-        return;
-    }
-    if (unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-    if (!unitTarget->IsInWorld())
-    {
-        return;
-    }
-    if (m_spellInfo->Id == 22012 && !unitTarget->HasAura(2584))
-    {
-        return;
-    }
-
-    ((Player*)unitTarget)->ResurrectPlayer(1.0f);
-    ((Player*)unitTarget)->SpawnCorpseBones();
-}
-
-// remove insignia spell effect
-void Spell::EffectSkinPlayerCorpse(SpellEffectIndex /*eff_idx*/)
-{
-    DEBUG_LOG("Effect: SkinPlayerCorpse");
-    if ((m_caster->GetTypeId() != TYPEID_PLAYER) || (unitTarget->GetTypeId() != TYPEID_PLAYER) || (unitTarget->IsAlive()))
-    {
-        return;
-    }
-
-    ((Player*)unitTarget)->RemovedInsignia((Player*)m_caster);
-}
-
-void Spell::EffectStealBeneficialBuff(SpellEffectIndex eff_idx)
-{
-    DEBUG_LOG("Effect: StealBeneficialBuff");
-
-    if (!unitTarget || unitTarget == m_caster)              // can't steal from self
-    {
-        return;
-    }
-
-    typedef std::vector<SpellAuraHolder*> StealList;
-    StealList steal_list;
-    // Create dispel mask by dispel type
-    uint32 dispelMask  = GetDispellMask(DispelType(m_spellInfo->EffectMiscValue[eff_idx]));
-    Unit::SpellAuraHolderMap const& auras = unitTarget->GetSpellAuraHolderMap();
-    for (Unit::SpellAuraHolderMap::const_iterator itr = auras.begin(); itr != auras.end(); ++itr)
-    {
-        SpellAuraHolder* holder = itr->second;
-        if (holder && (1 << holder->GetSpellProto()->Dispel) & dispelMask)
-        {
-            // Need check for passive? this
-            if (holder->IsPositive() && !holder->IsPassive() && !holder->GetSpellProto()->HasAttribute(SPELL_ATTR_EX4_NOT_STEALABLE))
-            {
-                steal_list.push_back(holder);
-            }
-        }
-    }
-    // Ok if exist some buffs for dispel try dispel it
-    if (!steal_list.empty())
-    {
-        typedef std::list < std::pair<uint32, ObjectGuid> > SuccessList;
-        SuccessList success_list;
-        int32 list_size = steal_list.size();
-        // Dispell N = damage buffs (or while exist buffs for dispel)
-        for (int32 count = 0; count < damage && list_size > 0; ++count)
-        {
-            // Random select buff for dispel
-            SpellAuraHolder* holder = steal_list[urand(0, list_size-1)];
-            // Not use chance for steal
-            // TODO possible need do it
-            success_list.push_back(SuccessList::value_type(holder->GetId(), holder->GetCasterGuid()));
-
-            // Remove buff from list for prevent doubles
-            for (StealList::iterator j = steal_list.begin(); j != steal_list.end();)
-            {
-                SpellAuraHolder* stealed = *j;
-                if (stealed->GetId() == holder->GetId() && stealed->GetCasterGuid() == holder->GetCasterGuid())
-                {
-                    j = steal_list.erase(j);
-                    --list_size;
-                }
-                else
-                {
-                    ++j;
-                }
-            }
-        }
-        // Really try steal and send log
-        if (!success_list.empty())
-        {
-            int32 count = success_list.size();
-            WorldPacket data(SMSG_SPELLSTEALLOG, 8 + 8 + 4 + 1 + 4 + count * 5);
-            data << unitTarget->GetPackGUID();       // Victim GUID
-            data << m_caster->GetPackGUID();         // Caster GUID
-            data << uint32(m_spellInfo->Id);         // Dispell spell id
-            data << uint8(0);                        // not used
-            data << uint32(count);                   // count
-            for (SuccessList::iterator j = success_list.begin(); j != success_list.end(); ++j)
-            {
-                SpellEntry const* spellInfo = sSpellStore.LookupEntry(j->first);
-                data << uint32(spellInfo->Id);       // Spell Id
-                data << uint8(0);                    // 0 - steals !=0 transfers
-                unitTarget->RemoveAurasDueToSpellBySteal(spellInfo->Id, j->second, m_caster);
-            }
-            m_caster->SendMessageToSet(&data, true);
-        }
-    }
-}
-
-void Spell::EffectKillCreditGroup(SpellEffectIndex eff_idx)
-{
-    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    ((Player*)unitTarget)->RewardPlayerAndGroupAtEvent(m_spellInfo->EffectMiscValue[eff_idx], unitTarget);
-}
-
-void Spell::EffectQuestFail(SpellEffectIndex eff_idx)
-{
-    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    ((Player*)unitTarget)->FailQuest(m_spellInfo->EffectMiscValue[eff_idx]);
-}
-
-void Spell::EffectPlaySound(SpellEffectIndex eff_idx)
-{
-    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    uint32 soundId = m_spellInfo->EffectMiscValue[eff_idx];
-    if (!sSoundEntriesStore.LookupEntry(soundId))
-    {
-        sLog.outError("EffectPlaySound: Sound (Id: %u) in spell %u does not exist.", soundId, m_spellInfo->Id);
-        return;
-    }
-
-    unitTarget->PlayDirectSound(soundId, (Player*)unitTarget);
-}
-
-void Spell::EffectPlayMusic(SpellEffectIndex eff_idx)
-{
-    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    uint32 soundId = m_spellInfo->EffectMiscValue[eff_idx];
-    if (!sSoundEntriesStore.LookupEntry(soundId))
-    {
-        sLog.outError("EffectPlayMusic: Sound (Id: %u) in spell %u does not exist.", soundId, m_spellInfo->Id);
-        return;
-    }
-
-    m_caster->PlayMusic(soundId, (Player*)unitTarget);
-}
-
-/**
- * @brief Sets the player's homebind location to the current position.
- *
- * @param eff_idx The bind effect index.
- */
-void Spell::EffectBind(SpellEffectIndex eff_idx)
-{
-    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    Player* player = (Player*)unitTarget;
-
-    uint32 area_id;
-    WorldLocation loc;
-    player->GetPosition(loc);
-    area_id = player->GetAreaId();
-
-    player->SetHomebindToLocation(loc, area_id);
-
-    // binding
-    WorldPacket data(SMSG_BINDPOINTUPDATE, (4 + 4 + 4 + 4 + 4));
-    data << float(loc.coord_x);
-    data << float(loc.coord_y);
-    data << float(loc.coord_z);
-    data << uint32(loc.mapid);
-    data << uint32(area_id);
-    player->SendDirectMessage(&data);
-
-    DEBUG_LOG("New Home Position X is %f", loc.coord_x);
-    DEBUG_LOG("New Home Position Y is %f", loc.coord_y);
-    DEBUG_LOG("New Home Position Z is %f", loc.coord_z);
-    DEBUG_LOG("New Home MapId is %u", loc.mapid);
-    DEBUG_LOG("New Home AreaId is %u", area_id);
-
-    // zone update
-    data.Initialize(SMSG_PLAYERBOUND, 8 + 4);
-    data << m_caster->GetObjectGuid();
-    data << uint32(area_id);
-    player->SendDirectMessage(&data);
-}
-
-/**
- * @brief Sends a battleground player target to its graveyard.
- */
-void Spell::EffectRedirectThreat(SpellEffectIndex /*eff_idx*/)
-{
-    if (!unitTarget)
-    {
-        return;
-    }
-
-    m_caster->GetHostileRefManager().SetThreatRedirection(unitTarget->GetObjectGuid());
-}
-
-void Spell::EffectKnockBackFromPosition(SpellEffectIndex eff_idx)
-{
-    if (!unitTarget || unitTarget->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    float x, y, z;
-    if (m_targets.m_targetMask & TARGET_FLAG_DEST_LOCATION)
-    {
-        m_targets.getDestination(x, y, z);
-    }
-    else
-    {
-        m_caster->GetPosition(x, y, z);
-    }
-
-    float angle = unitTarget->GetAngle(x, y) + M_PI_F;
-    float horizontalSpeed = m_spellInfo->EffectMiscValue[eff_idx] * 0.1f;
-    float verticalSpeed = damage * 0.1f;
-    ((Player*)unitTarget)->GetSession()->SendKnockBack(angle, horizontalSpeed, verticalSpeed);
-}
